requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-mail==0.9.1
opencv-python==4.5.3.56
pandas==1.3.3
openpyxl==3.0.9
python-dotenv==0.19.0
easyocr==1.4.1
numpy==1.21.2